Add files via upload
Login and sign up are good. Have create chat making a chatsession and a messages collection within that chatsession. For my chats i have the messages being recieved and showing on the list view. IM not sure how replies work i didnt get that far. For the create chat screen i did the users wrong, we need to set a custom adapter to set eh list view i thought we didnt so i just used a regular one. I also have no idea how to check if someone is online. Another thing is i need to get the name to place it in the mychats fragments. Ill hop on later to do some more.